requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

Flask-WTF==1.2.2
Flask-CORS==5.0.1
Flask-Bcrypt==1.0.1
libsql-client>=0.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
email-validator==2.2.0
